requirements.txt
aiohttp==3.8.5
aiomysql==0.2.0
APScheduler==3.10.4
asyncpg==0.30.0
celery==5.4.0
httpx==0.28.1
nest_asyncio==1.5.6
openai==1.65.2
PyMySQL==1.1.1
pytest==8.3.3
python-dotenv==1.0.1
aiolimiter==1.1.0
python-telegram-bot[rate-limiter]==21.11.1
python_bcrypt==0.3.2
sentry_sdk==2.18.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.0.0+cpu
transformers==4.46.3
cryptography==43.0.0
